Can now view up to Risks page but can save properly yet

+++ ModernSlavery.Core.Entities/Statement.cs
@@ -167,9 +167,9 @@ namespace ModernSlavery.Core.Entities
 
         public string KeyAchievements { get; set; }
 
-        public decimal MinStatementYears { get; set; }
+        public byte MinStatementYears { get; set; }
 
-        public decimal? MaxStatementYears { get; set; }
+        public byte? MaxStatementYears { get; set; }
 
         #endregion
